requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data APIs
alpaca-py>=0.20.0
alpaca-trade-api>=3.0.0
python-binance>=1.0.19
ccxt>=4.0.0
yfinance>=0.2.28

# Technical Analysis
ta>=0.10.2

# Database
sqlalchemy>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0
pyyaml>=6.0
pyarrow>=12.0.0
fastparquet>=2023.4.0

# API/Web
requests>=2.31.0
aiohttp>=3.8.0
websockets>=9.0,<11
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Jupyter (optional for exploration)
jupyter>=1.0.0
ipywidgets>=8.0.0
